+++ Source/Podio .NET/Models/AppMarketShares.cs
@@ -6,7 +6,7 @@ namespace PodioAPI.Models
     public class AppMarketShares
     {
         [JsonProperty("total")]
-        public int? Total { get; set; }
+        public long? Total { get; set; }
 
         [JsonProperty("shares")]
         public List<AppMarketShare> Shares { get; set; }
