+++ Source/Podio .NET/Models/FieldConfig.cs
@@ -39,7 +39,7 @@ namespace PodioAPI.Models
         ///     An integer indicating the order of the field compared to other fields
         /// </summary>
         [JsonProperty("delta", NullValueHandling = NullValueHandling.Ignore)]
-        public int? Delta { get; set; }
+        public long? Delta { get; set; }
 
         [JsonProperty("hidden", NullValueHandling = NullValueHandling.Ignore)]
         public bool? Hidden { get; set; }
