+++ Source/Podio .NET/Models/ImporterInfo.cs
@@ -6,7 +6,7 @@ namespace PodioAPI.Models
     public class ImporterInfo
     {
         [JsonProperty("row_count")]
-        public int RowCount { get; set; }
+        public long RowCount { get; set; }
 
         [JsonProperty("columns")]
         private List<FileColumn> Columns { get; set; }
@@ -15,7 +15,7 @@ namespace PodioAPI.Models
     public class FileColumn
     {
         [JsonProperty("id")]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [JsonProperty("name")]
         public string Name { get; set; }
