+++ Source/Podio .NET/Models/ItemMicro.cs
@@ -5,10 +5,10 @@ namespace PodioAPI.Models
     public class ItemMicro
     {
         [JsonProperty("app_item_id")]
-        public int AppItemId { get; set; }
+        public long AppItemId { get; set; }
 
         [JsonProperty("item_id")]
-        public int ItemId { get; set; }
+        public long ItemId { get; set; }
 
         [JsonProperty("title")]
         public string Title { get; set; }
@@ -17,6 +17,6 @@ namespace PodioAPI.Models
         public string Link { get; set; }
 
         [JsonProperty("revision")]
-        public int Revision { get; set; }
+        public long Revision { get; set; }
     }
 }
