+++ Source/Podio .NET/Models/Profile.cs
@@ -6,10 +6,10 @@ namespace PodioAPI.Models
     public class Profile
     {
         [JsonProperty("profile_id")]
-        public int ProfileId { get; set; }
+        public long ProfileId { get; set; }
 
         [JsonProperty("avatar")]
-        public int? Avatar { get; set; }
+        public long? Avatar { get; set; }
 
         [JsonProperty("image")]
         public Image Image { get; set; }
@@ -24,15 +24,15 @@ namespace PodioAPI.Models
         public string Name { get; set; }
 
         [JsonProperty("org_id")]
-        public int? OrgId { get; set; }
+        public long? OrgId { get; set; }
 
         [JsonProperty("space_id")]
-        public int? SpaceId { get; set; }
+        public long? SpaceId { get; set; }
 
         [JsonProperty("type")]
         public string Type { get; set; }
 
         [JsonProperty("user_id")]
-        public int? UserId { get; set; }
+        public long? UserId { get; set; }
     }
 }
